menambahkan kriteria edit

+++ src/main/java/com/gstudio/saw/repository/BobotRepository.java
@@ -3,5 +3,5 @@ package com.gstudio.saw.repository;
 import com.gstudio.saw.model.Bobot;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BobotRepository extends CrudRepository<Bobot, Float> {
+public interface BobotRepository extends CrudRepository<Bobot, Integer> {
 }
